feature: toss payment api

+++ domain/src/main/kotlin/com/example/domain/model/product/Inventory.kt
@@ -5,7 +5,7 @@ import java.time.Instant
 class Inventory(
     var id: Long? = null,
     var productId: Long? = null,
-    var quantity: Int? = null,
+    var quantity: Long? = null,
     var createdDt: Instant? = null,
     var updatedDt: Instant? = null,
 ) {
