+++ sql/product/1_productInit.sql
@@ -12,7 +12,7 @@ create table `products` (
 create table `inventory` (
     `id`                bigint NOT NULL AUTO_INCREMENT,
     `product_id`        bigint,
-    `quantity`          INTEGER,
+    `quantity`          bigint,
     `created_dt` timestamp   not null default current_timestamp,
     `updated_dt` timestamp   not null default current_timestamp on update current_timestamp,
     PRIMARY KEY (`id`)
